Fixing test leak with system heap in hybrid heap test. This was not caught locally by the tools (odd) Fixing Windows build issue with missing header Fixing some warnings with unreferenced params that were only triggered by Windows VC compiler Fixing some warnings with type cast loss of precision warnings triggered by Windows VC compiler

+++ src/client/src/StreamEvent.c
@@ -966,9 +966,9 @@ STATUS calculateCallLatency(PKinesisVideoStream pKinesisVideoStream, BOOL cplApi
 
     // Exponential mean averaging
     if (cplApiCall) {
-        pKinesisVideoStream->diagnostics.cplApiCallLatency = EMA_ACCUMULATOR_GET_NEXT(pKinesisVideoStream->diagnostics.cplApiCallLatency, latency);
+        pKinesisVideoStream->diagnostics.cplApiCallLatency = (UINT64) EMA_ACCUMULATOR_GET_NEXT(pKinesisVideoStream->diagnostics.cplApiCallLatency, latency);
     } else {
-        pKinesisVideoStream->diagnostics.dataApiCallLatency = EMA_ACCUMULATOR_GET_NEXT(pKinesisVideoStream->diagnostics.dataApiCallLatency, latency);
+        pKinesisVideoStream->diagnostics.dataApiCallLatency = (UINT64) EMA_ACCUMULATOR_GET_NEXT(pKinesisVideoStream->diagnostics.dataApiCallLatency, latency);
     }
 
 CleanUp:
